add redis and bigdecimal on variables

+++ src/main/java/com/danielsilva/imcApplication/dtos/ClienteDtoRequest.java
@@ -6,14 +6,16 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class ClienteDtoRequest {
                 String nome;
-                Double altura;
-                Double peso ;
+                BigDecimal altura;
+                BigDecimal peso ;
                 String email ;
 
 }
